STEP 4.COMMIT 2021-12-25
- FeedRepository 추상클래스 삭제. FeedRepositoryJpa로 대체FeedRepository. @PersistenceContext 를 이용한 EntityManagerFactory, EntityManger 컨트롤
- HomeController 삭제. index.html 을 기본 template
- yml 설정 수정. sql 및 parameter logging 출력
- SpringConfig 삭제. @을 이용한 컴포넌트 컨트롤
