+++ src/main/resources/application.yml
@@ -2,12 +2,19 @@ spring:
   datasource:
     url: jdbc:mariadb://localhost:3306/dalee
     driverClassName: org.mariadb.jdbc.Driver
-    username: #
-    password: #
+    username:
+    password:
+
   jpa:
-    database-platform: org.hibernate.dialect.MySQL5InnoDBDialect
     hibernate:
       ddl-auto: create
-      show-sql: true
+    show-sql: true
+
 server:
   port: 8080
+
+logging:
+  level:
+    org.hibernate:
+      type.descriptor.sql: trace
+      SQL: debug
