--- src/main/resources/templates/home.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <title>홈</title>
-</head>
-<body>
-<a href="/feed/list">피드</a>
-</body>
-</html>
